# EHN: Speed up common use of convection maps