@SEC-VT 2022-09-02
I found a way to vectorize the Legendre polynomial function from scipy:
```python
def vec_lpmn(m, n, x):
    return special.lpmn(m, n, x)[0]
vec_lpmn = np.vectorize(vec_lpmn, signature='(),(),()->(k,k)')
```
When creating numpy ufunc that return nonscalar outputs, you have to modify the [ufunc signature](https://numpy.org/doc/stable/reference/c-api/generalized-ufuncs.html#details-of-signature).

In my testing this doesn't speed up single plot generation (since instantiating a matplotlib axes itself takes about 4s), but there may be gains when updating the data on subsequent plots.

Elsewise, this PR looks good to me.