@carleyjmartin 2022-09-02
Thanks @SEC-VT !
If you think the PR is good, I think it only really needs 1 approval so feel free to merge. 
(We can stick the vectorizing in another PR when it's been put in later, smaller PRs are far easier to get merged, it might be good to make an issue on this actually)